fix: don't use floating point values in loops, found via SpotBugs
Do not use floating-point variables as loop counters
Using floating-point variables should not be used as loop counters, as they are not precise, which may result in incorrect loops. A loop counter is a variable that is changed with each iteration and controls when the loop should terminate. It is decreased or increased by a fixed amount each iteration.
See rule NUM09-J.

Bug kind and pattern: FL - FL_FLOATS_AS_LOOP_COUNTERS