Restored Logic for Facilities to Reveal StratCon Tracks
Restored the `revealTrack` property in facilities and integrated it into the StratCon logic to determine visibility. Restored drawing and visibility checks accordingly to account for facilities that enable track-wide reveals.

This ensures we have two facility modifiers relating to hex reveal: Scan Range (which increases line of sight), and Reveal Track (which reveals the whole Sector). The latter is not currently used by any Facilities, but might be useful for player custom facilities.