Added 'None' Generation Method to `ScenarioForceTemplate`
- Implemented a 'None' option in `ForceGenerationMethod` to allow bot forces where no units are generated, enabling more flexible scenario templates.
- Adjusted `AtBDynamicScenarioFactory` to account for this method in force generation logic.

This resolves a situation where we need to have all objective-significant Bot Forces generated when the scenario is created. However, if we want the Bot Force to have specific units and not random ones, we then need to remove the already generated ones. By using this method, we're able to generate an empty force during scenario creation which can then be populated later.

This will be used by the upcoming prisoner rework.