Fix #6794: Significantly Improved TO&E Reading Speed by Removing C3 Calculations
Fix #6794

We previously received feedback from a user that C3 was not being factored into BV calculations in the TO&E tab. We then went in and fixed this, however that caused the every C3 equipped unit needing to read the ammo loadout of every unit in the TO&E any time the TO&E was interacted with. This resulted in a lock-up for large campaigns.