Added 'Advanced Multiple Days' & 'Mass Personnel Training' Buttons
### Dev Notes
With all the traveling campaigns go through it didn't make sense to have the advance multiple days option hidden behind a menu.

Similarly, we see plenty of users unaware that the mass training dialog exists (until they come onto the Discord asking for this feature to be added). So moving it out of a menu made a lot of sense.