Improvement: Added Tooltips that Explain the Skill When Viewing a Person in the Personnel Tab or Purchasing/Improving Skills
### Dev Notes
We frequently get contacted by users querying what one skill or another does. With the introduction of both aging effects, SPAs that affect skills, traits, _and_ attribute scores this was only going to get worse.

I went ahead and created a set of tooltips that should significantly reduce confusion. Now, whenever a player is purchasing (or improving) a skill, or viewing a character in Person View, simply hovering over the skill will show not just what the skill does but what modifiers are affecting it.